+++ src/main/java/com/ureca/snac/auth/controller/verify/SnsController.java
@@ -4,7 +4,9 @@ import com.ureca.snac.auth.dto.request.PhoneRequest;
 import com.ureca.snac.auth.dto.request.VerificationPhoneRequest;
 import com.ureca.snac.auth.service.verify.SnsService;
 import com.ureca.snac.common.ApiResponse;
+import com.ureca.snac.config.RabbitMQConfig;
 import lombok.RequiredArgsConstructor;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,10 +18,12 @@ import static com.ureca.snac.common.BaseCode.*;
 public class SnsController implements SnsControllerSwagger {
 
     private final SnsService snsService;
+    private final RabbitTemplate rabbitTemplate;
 
     @Override
     public ResponseEntity<ApiResponse<Void>> sendVerificationCode(@RequestBody PhoneRequest dto) {
-        snsService.sendVerificationCode(dto.getPhone());
+        rabbitTemplate.convertAndSend(RabbitMQConfig.SMS_EXCHANGE, RabbitMQConfig.SMS_AUTH_ROUTING_KEY, dto.getPhone());
+//        snsService.sendVerificationCode(dto.getPhone()); // 기존 코드를 rabbitMQ 비동기로 변경했습니다.
         return ResponseEntity.ok(ApiResponse.ok(SMS_VERIFICATION_SENT));
     }
