+++ src/main/java/com/ureca/snac/auth/service/verify/SnsService.java
@@ -2,7 +2,7 @@ package com.ureca.snac.auth.service.verify;
 
 public interface SnsService {
 
-    void sendVerificationCode(String phoneNumber);
+//    void sendVerificationCode(String phoneNumber);
 
     void verifyCode(String phoneNumber, String code);
